Fixed decimal places in baking

+++ database/migrations/2022_08_17_122140_create_accounts_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string("bank_name");
             $table->string("branch");
             $table->string("branch_code");
-            $table->integer("balance");
+            $table->decimal("balance",8,2);
             $table->timestamps();
             $table->softDeletes();
         });
